move zscore functons to stats; any confidence interval can be entered
