+++ climpred/stats.py
@@ -22,6 +22,12 @@ Time Series
 `xr_autocorr` : Calculates the autocorrelation of time series over some lag.
 `xr_tau_d` : Calculates the decorrelation time of a time series.
 
+Z Scores
+--------
+`_z_score`: Returns the z score at a particular confidence.
+`z_significance`: Computes statistical significance between two correlation
+coefficients using Fisher's r to z conversion.
+
 Generate Time Series Data
 -------------------------
 `create_power_spectrum` : Creates power spectrum with CI for a given pd.series.
@@ -32,7 +38,7 @@ import pandas as pd
 import scipy.stats as ss
 import xarray as xr
 from scipy.signal import detrend, periodogram, tukey
-from scipy.stats import chi2, linregress
+from scipy.stats import chi2, linregress, norm
 from xskillscore import pearson_r, pearson_r_p_value
 
 
@@ -390,7 +396,7 @@ def xr_rm_poly(ds, order, dim='time'):
 
     Input
     -----
-    ds : xarray object 
+    ds : xarray object
         Single time series or many gridded time series of object to be
         detrended
     order : int
@@ -401,7 +407,7 @@ def xr_rm_poly(ds, order, dim='time'):
 
     Returns
     -------
-    detrended_ts : xarray object 
+    detrended_ts : xarray object
         DataArray or Dataset with detrended time series.
     """
     _check_xarray(ds)
@@ -437,8 +443,8 @@ def xr_rm_poly(ds, order, dim='time'):
         new_ds = xr.Dataset()
         for i, var in enumerate(store_vars):
             new_ds[var] = xr.DataArray(y.isel(variable=i))
-        return new_ds 
-    
+        return new_ds
+
     if isinstance(ds, xr.Dataset):
         DATASET, store_vars = True, ds.data_vars
         y = []
@@ -463,7 +469,7 @@ def xr_rm_poly(ds, order, dim='time'):
     detrended_ds = xr.DataArray(detrended_ts, dims=dims, coords=coords)
     if DATASET:
         detrended_ds = _reconstruct_ds(detrended_ds, store_vars)
-    return detrended_ds 
+    return detrended_ds
 
 
 def xr_rm_trend(da, dim='time'):
@@ -569,6 +575,51 @@ def xr_decorrelation_time(da, r=20, dim='time'):
     one = da.mean(dim) / da.mean(dim)
     return one + 2 * xr.concat([xr_autocorr(da, dim=dim, lag=i) ** i for i in
                                 range(1, r)], 'it').sum('it')
+
+
+# -------
+# Z SCORE
+# -------
+def _z_score(ci):
+    """Returns critical z score given a confidence interval
+
+    Source: https://stackoverflow.com/questions/20864847/
+            probability-to-z-score-and-vice-versa-in-python
+    """
+    diff = (100 - ci) / 2
+    return norm.ppf((100 - diff) / 100)
+
+
+def z_significance(r1, r2, N, ci=90):
+    """Computes the z test statistic for two ACC time series, e.g. an
+       initialized ensemble ACC and persistence forecast ACC.
+
+    Inputs:
+        r1, r2: (xarray objects) time series, grids, etc. of pearson
+                correlation coefficients between the two prediction systems
+                of interest.
+        N: (int) length of original time series being correlated.
+        ci: (optional int) confidence level for z-statistic test
+
+    Returns:
+        Boolean array of same dimensions as input where True means r1 is
+        significantly different from r2 at ci.
+
+    Reference:
+        https://www.statisticssolutions.com/comparing-correlation-coefficients/
+    """
+    def _r_to_z(r):
+        """Fisher's r to z transformation"""
+        return 0.5 * (np.log(1 + r) - np.log(1 - r))
+
+    z1, z2 = _r_to_z(r1), _r_to_z(r2)
+    difference = np.abs(z1 - z2)
+    zo = difference / (np.sqrt(2*(1 / (N - 3))))
+    confidence = np.zeros_like(zo)
+    confidence[:] = _z_score(ci)
+    sig = xr.DataArray(zo > confidence)
+    return sig
+
 # --------------------------------------------#
 # GENERATE TIME SERIES DATA
 # Functions that create time series data
